pyproject.toml
[project]
name = "telebot"
version = "0.1.0"
description = ""
authors = [
    {name = "AngeloGagno",email = "angagelogagno@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "pytelegrambotapi (>=4.27.0,<5.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "tinydb (>=4.8.2,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
